requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
